macros-tracker/scripts/requirements.txt
opencv-python>=4.7.0
numpy>=1.24.0
pillow>=9.5.0
transformers>=4.30.0
torch>=2.0.0
mistralai>=0.0.7
requests>=2.31.0 